Added standard CDK tooling for use across our apps (#18)
* feat: added support for Localstack including `NodeJSFunction` generator
* chore: updated readme
* feat: automatically instrument Datadog
* chore: switch from logRetention to explicit logGroup, set retention to 1 year
