+++ package.json
@@ -9,8 +9,9 @@
   "author": "Nyman Turkish",
   "license": "ISC",
   "dependencies": {
-    "aws-cdk-lib": "^2.124.0",
-    "aws-lambda": "^1.0.7"
+    "aws-cdk-lib": "^2.150.0",
+    "aws-lambda": "^1.0.7",
+    "datadog-cdk-constructs-v2": "^1.16.1"
   },
   "devDependencies": {
     "@types/node": "^20.11.10",
